requirements.txt
fastapi==0.115.8
uvicorn==0.33.0
transformers==4.46.3
torch==2.4.1
psycopg2-binary==2.9.5
python-dotenv==1.0.1
google-genai==1.2.0
translate==3.6.1
SQLAlchemy==2.0.30
ipython==8.18.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt=4.0.1
python-multipart==0.0.6
alembic==1.16.5